requirements.txt
fastapi
uvicorn
paho-mqtt
tensorflow
numpy
